+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:intl_phone_field_plus/intl_phone_field.dart';
+import 'package:intl_phone_field_continued/intl_phone_field.dart';
 
 void main() {
   runApp(const MyApp());